Add DOT to double rewards, add assets.mint & assets.burn history (#468)
* update predefined assets & add DOT to double price pool

* add asset burn & mint history elements
